Heroku install

+++ api/controllers/categories.js
@@ -1,5 +1,6 @@
 import db from '../models/connect';
 import validateCategories from '../validation/categories';
+import { checkToken, adminAccess, userAccess } from '../validation/auth';
 
 const Category = {
   /**
@@ -16,21 +17,27 @@ const Category = {
     if (!isValid) {
       return res.status(400).json(errors);
     }
-    const text = `INSERT INTO
+    let noAccess = false;
+    checkToken(req, res);
+    noAccess = true;
+    adminAccess(req, res);
+    if (!noAccess) {
+      const text = `INSERT INTO
       categories(categoryName,categoryDetails)
       VALUES($1, $2)
       returning *`;
-    const values = [
-      req.body.categoryName,
-      req.body.categoryDetails,
-    ];
+      const values = [
+        req.body.categoryName,
+        req.body.categoryDetails,
+      ];
 
-    try {
-      const { rows } = await db.query(text, values);
-      return res.status(201).json(rows[0]);
-    } catch (error) {
-      console.log(error);
-      return res.status(400).json(error);
+      try {
+        const { rows } = await db.query(text, values);
+        return res.status(201).json(rows[0]);
+      } catch (error) {
+        console.log(error);
+        return res.status(400).json(error);
+      }
     }
   },
   /**
@@ -81,25 +88,31 @@ const Category = {
     if (!isValid) {
       return res.status(400).json(errors);
     }
-    const findOneQuery = 'SELECT * FROM categories WHERE id=$1';
-    const updateOneQuery = `UPDATE categories
+    let noAccess = false;
+    checkToken(req, res);
+    noAccess = true;
+    adminAccess(req, res);
+    if (!noAccess) {
+      const findOneQuery = 'SELECT * FROM categories WHERE id=$1';
+      const updateOneQuery = `UPDATE categories
       SET categoryName=$1,categoryDetails=$2
       WHERE id=$3 returning *`;
-    try {
-      const { rows } = await db.query(findOneQuery, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).json({ message: 'category not found' });
+      try {
+        const { rows } = await db.query(findOneQuery, [req.params.id]);
+        if (!rows[0]) {
+          return res.status(404).json({ message: 'category not found' });
+        }
+        const values = [
+          req.body.categoryName || rows[0].categoryName,
+          req.body.categoryDetails || rows[0].categoryDetails,
+          req.params.id,
+        ];
+        const response = await db.query(updateOneQuery, values);
+        return res.status(200).json(response.rows[0]);
+      } catch (err) {
+        console.log(err);
+        return res.status(400).json(err);
       }
-      const values = [
-        req.body.categoryName || rows[0].categoryName,
-        req.body.categoryDetails || rows[0].categoryDetails,
-        req.params.id,
-      ];
-      const response = await db.query(updateOneQuery, values);
-      return res.status(200).json(response.rows[0]);
-    } catch (err) {
-      console.log(err);
-      return res.status(400).json(err);
     }
   },
   /**
@@ -109,16 +122,22 @@ const Category = {
    * @returns {void} return statuc code 204
    */
   async delete(req, res) {
-    const deleteQuery = `DELETE FROM categories WHERE id=${req.params.id} returning *`;
-    try {
-      const { rows } = await db.query(deleteQuery);
-      if (!rows[0]) {
-        return res.status(404).json({ message: 'category not found' });
+    let noAccess = false;
+    checkToken(req, res);
+    noAccess = true;
+    adminAccess(req, res);
+    if (!noAccess) {
+      const deleteQuery = `DELETE FROM categories WHERE id=${req.params.id} returning *`;
+      try {
+        const { rows } = await db.query(deleteQuery);
+        if (!rows[0]) {
+          return res.status(404).json({ message: 'category not found' });
+        }
+        return res.status(204).json({ message: 'deleted' });
+      } catch (error) {
+        console.log(error);
+        return res.status(400).json(error);
       }
-      return res.status(204).json({ message: 'deleted' });
-    } catch (error) {
-      console.log(error);
-      return res.status(400).json(error);
     }
   },
 };
